template_lib/requirements.txt
torchviz
easydict
graphviz
gym
imageio
matplotlib
numpy
Pillow
scikit-image
scikit-learn
scipy
tensorboardX==1.7
tensorflow==1.13.1
torch
torchvision
tqdm
pyyaml
gitpython